+++ src/core/commands/cluster/add.rs
@@ -2,12 +2,18 @@ use std::path::PathBuf;
 
 use anyhow::Result;
 
-use crate::core::Ctx;
+use crate::{api::model::Server, core::Ctx};
 
 pub(super) fn add(ctx: &mut Ctx, name: Option<String>, path: PathBuf) -> Result<()> {
     if let Some(c) = &mut ctx.cluster {
         let name = name.unwrap_or_else(|| path.file_name().unwrap().to_str().unwrap().to_string());
-        c.members.insert(name.clone(), path.canonicalize()?);
+        c.members.insert(
+            name.clone(),
+            Server {
+                path: path.canonicalize()?,
+                mod_dir: PathBuf::from("mods"),
+            },
+        );
         c.save()?;
         println!(
             "Added {}(at {}) to cluster {}",