+++ src/core/utils.rs
@@ -8,6 +8,7 @@ use crate::model::Mod;
 use anyhow::{anyhow, Context, Result};
 use directories::ProjectDirs;
 use log::debug;
+use log::info;
 use std::collections::HashMap;
 use std::ffi::OsStr;
 use std::fs::{self, File, OpenOptions};
@@ -32,16 +33,10 @@ pub async fn update_index(local: &Path, global: &Path) -> Vec<model::Mod> {
     let glob = LocalIndex::load(global);
     for e in index.iter_mut() {
         if let Ok(installed) = &installed {
-            e.installed = installed
-                .mods
-                .iter()
-                .any(|(n, f)| n == &e.name && f.version == e.version);
+            (*e).installed = installed.mods.iter().any(|(n, _)| n == &e.name);
         }
         if let Ok(glob) = &glob {
-            e.global = glob
-                .mods
-                .iter()
-                .any(|(n, f)| n == &e.name && f.version == e.version);
+            (*e).global = glob.mods.iter().any(|(n, _)| n == &e.name);
         }
     }
     println!(" Done!");